BatteryModel edited, slight changes to TransactionService.

+++ backend/src/main/java/com/backend/model/BatteryModel.java
@@ -19,8 +19,8 @@ public class BatteryModel {
     Integer id;
 
     String type;
-    Double recylePoint;
-    Double exchangePoint;
+    Float recylePoint;
+    Float exchangePoint;
     ZonedDateTime lastUpdate;
 
 }
